refactor rpc interface logic
